+++ app/Http/Livewire/Admin/NeighborhoodLivewire.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Livewire;
+namespace App\Http\Livewire\Admin;
 
 use App\Models\Neighborhood;
 use App\Models\Sector;
